+++ fabric/src/main/java/fr/loudo/narrativecraft/network/handlers/ServerPacketHandlerFabric.java
@@ -50,6 +50,7 @@ import fr.loudo.narrativecraft.network.story.C2SChoiceSelected;
 import fr.loudo.narrativecraft.network.story.C2SDialogueFinished;
 import fr.loudo.narrativecraft.network.story.C2SPlayStitchStory;
 import fr.loudo.narrativecraft.network.story.C2SPlayStory;
+import fr.loudo.narrativecraft.network.story.C2SSetStoryLocale;
 import fr.loudo.narrativecraft.network.story.C2SStopStory;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 
@@ -250,6 +251,13 @@ public class ServerPacketHandlerFabric {
                     C2SChangeGamemodePacket packet = C2SChangeGamemodePacket.read(buf);
                     server.execute(() -> {
                         ServerPacketHandler.changeGamemode(packet, player);
+                    });
+                });
+        ServerPlayNetworking.registerGlobalReceiver(
+                C2SSetStoryLocale.TYPE, (server, player, handler, buf, responseSender) -> {
+                    C2SSetStoryLocale packet = C2SSetStoryLocale.read(buf);
+                    server.execute(() -> {
+                        ServerPacketHandler.setStoryLocale(packet, player);
                     });
                 });
     }